fix : continued implementing and testing store-invitation
